+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed 
+
+- Github & Snowflake & OneDrive & Salesforce connectors can now be instantiated without backend fields or a secrets keeper
+
 ## [7.7.5] 2025-02-07
 
 ### Fixed